requirements.txt
flask
flask_session
pyrebase4
